Extract shared capabilities out of app/hiring
Modules imported by forms, education, partners, and core libs move to
neutral homes; hiring keeps only its domain logic:
- ChallengeBuilder → app/components/form-builder/FormBuilder
- ChallengeQuestionField → app/components/form-builder/QuestionField
  (symbol renamed FormQuestionField)
- info-body → app/components/form-builder/
- skills-rating, submission-check → app/lib/
- useChartColors → app/components/analytics/
- generic buildIcs → app/lib/ics.ts; interview-ics.ts keeps only the
  interview UID scheme + REQUEST/CANCEL wrappers

Cross-area imports that remain (cycles, confidentiality,
intern-eligibility, interview-emails) are hiring domain answers consumed
through its lib — the correct direction, left in place.
