+++ dali-api/app/routes/portal.apply.tsx
@@ -8,8 +8,8 @@ import { getApplicationsGmailRefreshToken } from "~/lib/gmail-integration";
 import { renderForSlot, notificationSlot } from "~/hiring/lib/email-variables";
 import { getActiveCycle } from "~/hiring/lib/cycles";
 import { reconcileDomainApplications } from "~/hiring/lib/domain-application";
-import { checkGitHubUrl, checkFigmaUrl, checkDriveUrl } from "~/hiring/lib/submission-check";
-import type { SubmissionCheckResult } from "~/hiring/lib/submission-check";
+import { checkGitHubUrl, checkFigmaUrl, checkDriveUrl } from "~/lib/submission-check";
+import type { SubmissionCheckResult } from "~/lib/submission-check";
 import { validateWordLimits } from "~/lib/word-count";
 import type { WordCountViolation } from "~/lib/word-count";
 import { isAnswered } from "~/lib/form-answers";
@@ -18,7 +18,7 @@ import { ApplicantErrorBoundary } from "~/components/ApplicantErrorBoundary";
 import { Modal } from "~/components/Modal";
 import { QuestionList } from "~/hiring/components/ApplicationAnswers";
 import { RichTextViewer, isEmptyDoc } from "~/components/RichTextViewer";
-import { type UrlCheckState } from "~/hiring/components/ChallengeQuestionField";
+import { type UrlCheckState } from "~/components/form-builder/QuestionField";
 import { FormField } from "~/forms/components/FormField";
 
 export const meta: Route.MetaFunction = () => [{ title: "Apply · DALI OS" }];
